requirements.txt
Eel==0.14.0
pyTelegramBotAPI==4.6.0
Pillow==9.0.1
requests==2.27.1
